requirements.txt
diffusers
matplotlib
numpy
opencv-python
pillow
torch
torchvision
transformers